updated with most recent changes
