--- cloud_security/iam_policyIds_4thFeb.csv
@@ -1,102 +1,0 @@
-PolicyIDs
-eeca101d-66b7-4e59-0007-3da5435da611
-e9a84ae4-1764-41f5-8565-5c848b3e0e67
-e231b07d-f1e1-40ad-9c7c-d38f09948db9
-99868a5a-9824-4da2-b413-ec5b1632b722
-5def0968-e071-46bb-a2b8-4a54f66bcf23
-59a1468f-150e-469b-a099-df1d372fbb92
-b6b547d3-e4cb-4420-83c2-a67f7a74d5f6
-9370db0b-6b0a-474f-8132-c8cd583a9356
-c85c9ae3-89c9-4b5c-ba7f-dc81d0d4b3ff
-49326617-2029-40c5-ac5f-e6ae939b2bdf
-882e73b6-6544-40ae-9993-bb2bfa3ec1d9
-85be5422-99bb-4e33-aefa-633885dc737d
-0c87cdbd-de14-47c8-b0c8-00703291bd29
-1aae9027-91c0-4696-a08d-cb6d7e390cce
-980343f0-8764-4f2c-8341-ffa15c9a5a0b
-19d9bbe7-81b8-4531-a8d6-c9e29474e18c
-7b0539b4-5fe4-4322-aff0-1535b24722db
-45dad4ec-29c4-4de4-be24-3043d385e209
-4eb8a151-91d2-4960-b645-6e953d0a79d7
-ab7d320a-0634-4c18-ac13-98b1f1f7c290
-ca4807f0-2fb8-4af5-949a-bd677b1501cf
-5c556a50-bbdf-4baa-bf0b-24a86ddb6bde
-309d1863-c8c7-4ff1-b7a8-454817c876d9
-4ed4f5d8-aa5a-4bb9-a91f-5684ac80ad57
-ceea9d4d-d15d-47c7-8f25-a84c8d78b50a
-7e64df2b-eeed-4b59-8b2e-55735d528e1a
-9cb7793b-f29e-4544-a9ba-451be9c3e3e8
-fa3d600e-101e-4e56-9961-6aa4baf693e4
-4b82cb4a-4d6c-4371-8501-a1a519e2ea11
-c14d0daf-bd4b-4b4f-8ee9-b41771217954
-385eee00-d051-41c6-9fb9-3e74c6c21be8
-1b4e2641-4315-4f41-a6da-e11ed8cd6990
-213a74d0-ad27-42af-bf35-2b34142177b9
-1579bf5e-2670-4d24-be3f-567f7195eb67
-f03a1ce8-67d6-4e43-a613-393756aa11dc
-b2678cef-2371-4129-9dc7-7e07669a4223
-216370af-1caf-4bcd-b39a-db7088830995
-b32f7071-4d0e-4bdb-bc96-a5774fcb41bd
-61a4e8d4-3f37-4733-bdf0-b7f1c41edb50
-cdb9aa21-d636-4dcb-ac46-58ac08d9490f
-c5e8d60e-108c-4383-983e-33b3dd468eff
-063eaecc-c487-40d9-a575-cf597645ee80
-6d3a90cd-8a90-4c98-bd9a-65f00f79d7fd
-7d0fb1b6-a64b-4ac8-90b8-42cfd737c258
-19c079bd-d022-47c2-b590-2ef41e76585e
-833d7175-b7ac-4b93-8fc1-ebb5440e7042
-0cc37adf-957f-4681-8043-88faee0515eb
-e6353326-77f7-4b5a-abdb-ede688ba3068
-ab9e5269-5014-48e6-b293-a318100aa535
-91e1a9f2-9bd4-4297-9b7f-64980658c97b
-8783eae9-6ef2-493a-ba92-104e33504976
-d6b9aaf4-2b05-45d7-9288-30232865136e
-228230b4-371a-4d75-9bce-10ffd586a446
-6c67217a-2f0f-4afd-b07e-17d640b300cc
-c6e9f36f-2a31-42bb-bf68-752b2f3c8e7b
-f13c6791-8e1d-4985-9473-8463c2ab3e95
-866ca1d4-40cb-4626-afa0-5a438a92019a
-cf87131d-0fa0-4f58-b4f1-6a401d89ee9f
-11ada836-c9b3-488c-809d-826c41a74a83
-c5023f4b-f36a-4a21-89fd-250106594635
-7a034f1e-bf01-4080-b9e1-9f1485b83508
-e9fb951f-4bbc-4045-87f8-c11af8eb0e46
-e174e94b-bc56-44a5-b1e8-5ad30e8cf039
-a6497b9e-4ee6-439a-8cba-86b0360bbe2f
-ad0685b6-49fc-4868-a424-b8bb8c962fd3
-83886acc-ec72-4898-8b1e-e19ab8b7afa8
-dd52863a-1820-4e5b-890a-8a084477c4a1
-5983d2c5-fd8a-44da-bf89-03519b9e6706
-9c36caff-cd53-4de7-810e-cc150906d910
-4b3891a0-9c6e-4231-9251-08d0b644e3eb
-32da183f-6263-4661-a2fd-fd18659013eb
-388202e0-e57b-44cb-ae81-e2d164a67b24
-c0a1c364-a2d1-432c-8a0f-b74ae4e09a3c
-1de8a24e-7a34-47b6-8de0-bcdfe97f1547
-1ca7610e-d95a-4a3a-abf4-104cdf5c43e8
-9c8f3ac1-e5d7-4457-bde9-463821881cc2
-a5d57ed3-55b9-4f4b-91ec-fcb6564f76ef
-568c9eb2-3f77-4cbd-93c4-955fc869265b
-7b3e8f84-6c60-411e-b6f9-c3182681fc24
-86ddf852-f3e7-4580-b323-e9569f1a2ddb
-101af573-5785-478e-b376-973522eb573b
-80642bcc-2903-4599-9a3f-c23905254ff9
-9226d312-f44b-42c6-bc68-c2ba3ab0ea92
-29afe858-4e43-4843-aa20-be46f4524136
-8de6c50e-3d48-495e-9f92-63c0c11e9e2a
-7a283e4c-0137-474c-a778-3cdb8fdd6ce3
-ec3f5960-dcbc-4704-ac37-ed1a03ff59b6
-50fea644-dd5e-4f1f-8e01-a36f4a66fee4
-7ee9fbb1-2e0c-47d2-b874-dea733148f04
-fe6f85ec-d339-416f-857a-370108c7f37c
-d713b788-5ba0-495f-b8bb-e530bec828e2
-69e15d91-44c9-47de-bdc2-df35001d35ec
-247d4e3a-ee78-4505-bcff-8e9d505f39aa
-8dfb8a24-6c2c-4b78-b7cc-314ccba0105a
-edf08e81-3014-4750-9b88-0a194dfe8a4e
-59f32bc9-5750-4821-b663-0c5b07ba3f85
-ae11f1c2-d147-4750-bcc5-e9862098b473
-c32418c8-9baf-45c0-a75c-9dd55f237dd2
-b9b7bb65-5bb0-46b4-92b6-5bb298e045b5
-04a8b74e-e5e3-4534-bc6b-b1bbf26fcc93
-2083569a-28b8-4e4b-a34d-222928277107
